make display name public to users & email private
Fullname renamed to display name. display name is visible to all other
system users.

Email is only visible to the user and admins.

Made wrapper classes for display name and email. Display name is
probably ok as is, email needs better validation of the address (see
TODO in class).

Updated Globus endpoint to only return email if the user is querying on
their own name

Also get rid of the session login stored in DB stuff. The plan is to
have a toggle on the first login page that'll set that.