Show disabled state & all remote usernames in UI
Instead of more or less picking a remote ID provider account randomly
and displaying that in the UI for any linked users, show all provider
usernames that are linked with the account.

Note the ID that is sent back to the server on account selection is
essentially random, but is not displayed to the user.

Show whether an account is disabled or not when picking accounts, and
disable the login button if so.