Don't throw errors from login redirect
After login at the 3rd party id provider, the 3rd party redirects to the
auth server which calls login(provider, authcode). The redirect url will
have an authcode in a query parameter.

If an error is thrown while the code is in this state, 1) the auth code
will be shown in the url for the error, and 2) any auth UI other than
the test ui contained in this repo can't catch and respond to the error.

As such, now only fatal errors are thrown in the login(provider,
authcode) method, since there's no way to continue and generally
speaking they shouldn't occur.

Now if a user's account is disabled, or non-admin logins (including
account creation) are disabled, that information is shown in the choice
page. It is the responsibility of the login(token, identityID) and
createUser(...) methods to check that login for a particular user is
allowed.