Avoid throwing errors from link redirect
After login at the 3rd party id provider, the 3rd party redirects to the
auth server which calls link(token, provider, authcode). The redirect
url will
have an authcode in a query parameter.

If an error is thrown while the code is in this state, 1) the auth code
will be shown in the url for the error, and 2) any auth UI other than
the test ui contained in this repo can't catch and respond to the error.

As such, now only fatal errors are thrown in the link(token, provider,
authcode) method, since there's no way to continue and generally
speaking they shouldn't occur.

Now if there are no available remote IDs to link, an exception is thrown
in the getLinkState() method.