Hacky fix for html/json selection
API was returning HTML in error response by default. This means that
code is html escaped by the templating engine, which causes the java
client to break because it expects '"' instead of '&quot' when it checks
the auth service url and expects an error.

Hack now returns json by default for /api/* and html by default for
other paths. Need to find a better solution but this'll work for now.