Add login endpoint allowing UI to intercept redirect from 3rd party
Adds a /login/complete/{provider} endpoint that assumes a UI element
intercepts the post-login redirect from a 3rd party identity provider,
and then makes an AJAX call with the state and authcode to the service.

The endpoint immediately sends back the same information that would
normally be returned from the /login/choice endpoint, except that when
possible (e.g. the remote identities are all associated with one user) a
token is returned for immediate login.