Add link endpoint allowing UI to intercept redirect from 3rd party
Adds a /link/complete/{provider} endpoint that assumes a UI element
intercepts the post-link redirect from a 3rd party identity provider,
and then makes an AJAX call with the state and authcode to the service.

The endpoint immediately sends back the same information that would
normally be returned from the /link/choice endpoint, except that when
possible (e.g. there's only one linkable identity) the link happens
immediately and that fact is returned in the body.