Add configuration option to never login a user immediately
Instead redirect to the UI and allow UI to handle the login. This is
useful for UIs that don't intercept the 3rd party redirect but want to
use the JSON endpoints that return the login token in the body rather
than in a cookie (or at least will in the future).