Remove JSON versions of /link/start and /login/start & remove token
They're useless. There's no way to have the redirect work correctly with
json input, has to be a form submit from the browser. There's no way to
set headers for a form submit. Worse, there's no way to catch any errors
for the form submit, which means the raw auth2 server response will be
shown if the token is invalid.

For the latter reason, removed the token check from /link/start.