Make RemoteIdentity local IDs fn(RemoteIdentity)
Assigning an arbitrary UUID to remote identities as a proxy for the
identity provider name and the id of the remote identity opens up a
class of bugs wherein IDs are swapped accidentally. Instead, create an
id that is a function of the provider name and the remote id.

Also fixed a bug with logging in with no policy ids.