Fix leaking password hashes
The LocalUser class returned from the database included a password hash
and salt, which is almost never needed. This class is returned from the
Authentication class and so that hash is exposed, although currently not
via the API. 

Split the hash and salt into a separate class and retrieval method and
only call that method when needed (e.g. for a login or on a password
change.

Also change the changePassword() storage method to take the same class.