Add an error storage system and use for provider errors.
Adds the ability to store an error type and error message with a
temporary token in the database instead of a set of identities. This
capability is used to store provider errors returned to the auth server
when the provider redirects to the /(login|link)/complete/<provider>
endpoint. The error is then returned when the client accesses any of
downstream link or login endpoints.

Currently only the provider error type is supported at the
Authentication.java layer. It should be relatively easy to add new error
types in future if needed.

Note that some error codes have changed - see the ErrorTypes file.