WIP on logging operation on token and role failures
Now logs when an operation fails because the user submitted an illegal
type of token (e.g. usually non-login). methods have been updated to
pass in the operation they're performing for login purposes when
checking tokens types and user roles.

Tests have been drastically refactored to use a common framework for
testing the token and user verification methods that are used by most
methods.

Contains WIP on logging failures due to a user not possessing a required
role.