Fix link hijack part 1
Make the following changes to the linking process to prevent link
hijacking. Link hijacking could occur if User A logs out while a link
process token is still active. If User B then logs in in the same
browser (even if it's been closed) User B could link User A's identities
to their account.

* /link/start now requires a login token, either in a cookie or posted
in the form.
* /link/start sets a temporary token in the database identifying the
user (currently lasts for 30m). This token is returned to the user as a
session cookie, so that if the browser is closed it's no longer
available.
* /link/complete now (effectively) adds the remote identities to the
temporary token and sets the token lifetime to 10m. The token is still
keyed to the specific user.
* /link/choice and /link/pick will now throw an error if the user
identified by the login token does not match that of the lnk process
token.
* delete temporary tokens from the DB when a login or link is complete.

Step 2: Refactor the temporary token code. It's way too complicated and
is waiting for bugs to be introduced. Add token types to prevent mixing
link start tokens, link choice tokens, and link complete tokens. Throw
an error is an inappropriate token is used at any of the steps.

Step 3: /logout should clear all link in process tokens for the user
(since login in process tokens don't belong to a specific user, they
cannot be cleared).