Refactor temporary token handling.
Made a builder for temporary session data to prevent another refactor
like this in the future.

Temp session data is now assigned an operation and that operation is
checked when accessing the token. This prevents using a stolen link
token to log in or vice versa, or rewriting the identities associated
with a link token.

The first stage of the linking process still assigns a temporary token
that identifies the user and is the sole identifier used in the second
stage. As such, this is set as a session cookie so it expires when the
browser is closed.

TODO: remove all temporary tokens associated with a user on logout.