Add test user integration test
Also contains a potential fix for a test error that so far only occurs
in TravisCI. It appears that locally the tests always call a particular
method out of the two methods for the KBase legacy POST endpoints, while
in TravisCI a random method is called. If the wrong method is called, an
unexpected error is thrown. Hence, the methods are combined and very
simplistic content negotiation is done at runtime rather than statically
via JAX-RS annotations.