Handle environments throughout the login and linking process.
Process flow for login/link (/ll):

* Client sends form to /ll/start with environment or null for default
environment.
* For login, if a user redirect url is sent it's checked against the
configured prefix for the appropriate environment.
* The environment is passed though the application to the appropriate
identity provider implementation to get the redirect url to the 3rd
party identity provider. The env-specific return redirect url is added
to the redirect url as a param.
* Redirect occurs to 3rd party. A cookie is set with the environment
name and another with the user redirect url, if any (login only).
* 3rd party redirects to the auth server with an auth token.
* The user redirect url cookie is checked as previously described.
* The environment and auth token are passed through the application to
the appropriate identity provider implementation to retrieve the user
identities. The environment is necessary so the implementation can pass
the correct return redirect url to the 3rd party identity provider,
which is required.
* If login/link are done at this point, all process tokens are nuked, a
login token is set (or a new identity is linked), and redirect occurs to
the user login redirect url, the configured final link redirect url for
the environment, or a default url.
* If not, the cookies are refreshed and redirect occurs to the
configured intermediate redirect url for the environment or a default
url.
* The user redirect url is checked again.
* The user then picks an account (login or link) or creates an account
(login only).
* The final check of the user redirect url occurs.
* All process tokens are nuked, a login token is set (or a new identity
is linked), and redirect occurs to the user login redirect url, the
configured final link redirect url for the environment, or a default
url.