Add a /testmode endpoint that mimics /
When starting services, it's good practice to ensure their dependecies
are up - e.g. when starting the workspace it checks that it can contact
auth.

The cheapest way to do this is to GET / on auth and check that the
response is as expected. However, since there is no corresponding
endpoint at /testmode, it's impossible to test this strategy. This PR
fixes that.